refactor(cli): restructure CLI code for better organization
- Gate `anyhow` dependency behind `cli` feature
- Create `src/cli/` module for CLI utilities (time, progress, output)
- Rename `src/bin/cmd/` to `src/bin/cmds/` for clarity
- Move `src/bin/common/` utilities into `src/cli/`

This change removes CLI dependencies from the core library,
providing a smaller dependency tree for library users.
